test: add background script state management tests
