+++ tests/test-helper.js
@@ -1,0 +1,98 @@
+const fs = require('node:fs')
+const vm = require('node:vm')
+const path = require('node:path')
+
+const bgScriptPath = path.join(__dirname, '..', 'background.js')
+const bgScriptContent = fs.readFileSync(bgScriptPath, 'utf8')
+
+function setupEnvironment(initialStorage = {}) {
+  const sessionSetData = []
+  let currentStorage = { ...initialStorage }
+
+  const chromeMock = {
+    storage: {
+      session: {
+        get: async (key) => {
+          return key ? { [key]: currentStorage[key] } : currentStorage
+        },
+        set: async (data) => {
+          sessionSetData.push(data)
+          currentStorage = { ...currentStorage, ...data }
+        }
+      },
+      sync: {
+        get: async () => ({})
+      },
+      local: {
+        get: async () => ({})
+      }
+    },
+    runtime: {
+      onMessage: { addListener: () => {} },
+      getPlatformInfo: async () => ({})
+    },
+    tabs: {
+      query: async () => [],
+      get: async (id) => ({ id, url: 'https://jules.google.com/u/0/session' }),
+      sendMessage: async () => ({
+        config: { at: 'token', bl: 'build', fsid: '123' },
+        accountNum: '0',
+        account: 'default'
+      })
+    },
+    scripting: {
+      executeScript: async () => {}
+    }
+  }
+
+  const sandbox = {
+    chrome: chromeMock,
+    fetch: async () => ({ ok: true, json: async () => [], text: async () => ")]}'\n\n4\n[[]]" }),
+    setTimeout,
+    setInterval,
+    clearInterval,
+    Math,
+    Date,
+    JSON,
+    String,
+    Array,
+    Map,
+    Object,
+    Error,
+    URLSearchParams,
+    Promise,
+    console,
+    parseInt
+  }
+
+  vm.createContext(sandbox)
+
+  const scriptContent =
+    bgScriptContent +
+    `\n
+    globalThis.test_stateReadyPromise = stateReadyPromise;
+    globalThis.test_state = () => state;
+    globalThis.test_updateState = updateState;
+    globalThis.test_addLog = addLog;
+    globalThis.test_buildBatchRequest = buildBatchRequest;
+    globalThis.test_fixJsonControlChars = fixJsonControlChars;
+    globalThis.test_findJsonEnd = findJsonEnd;
+    globalThis.test_parseResponse = parseResponse;
+    globalThis.test_parseTask = parseTask;
+    globalThis.test_TASK = TASK;
+    globalThis.test_parseSuggestion = parseSuggestion;
+    globalThis.test_buildSuggestionPrompt = buildSuggestionPrompt;
+    globalThis.test_buildStartPayload = buildStartPayload;
+    globalThis.test_SUGGESTION = SUGGESTION;
+    globalThis.test_SDETAIL = SDETAIL;
+    globalThis.test_CATEGORY_CONFIG = CATEGORY_CONFIG;
+    globalThis.test_DEFAULT_CATEGORY = DEFAULT_CATEGORY;
+  `
+
+  const script = new vm.Script(scriptContent)
+  script.runInContext(sandbox)
+
+  return { sandbox, sessionSetData }
+}
+
+module.exports = { setupEnvironment }
